refactor: replace Monaco theme scheduler with singleton service
The old monacoThemeScheduler stored a single setTheme function reference
from whichever component instance called last. When multiple code blocks
fired theme changes concurrently, they overwrote each other's setTheme,
causing Monaco to fall back to the default white theme.

New monacoThemeService maintains a registry of all mounted components'
setTheme functions. Theme changes pick any available function from the
registry, with proper deduplication. Components register on mount and
unregister on unmount, so stale references are never used.
